python3Packages.notmuch2: use bindingconfig that works on darwin
Instead of using the configure generated _notmuch_config.py, we generate
our own which correctly references each respective field.

The _notmuch_config.py in notmuch.bindingconfig contained references to
the build directory. On linux this didn't break anything due to the
sandbox always having a build root of /build. But on darwin this would
be different depending on the derivation being built. So the build
accidently worked since version.txt would end up in the same place on
linux.

It seems in practice NOTMUCH_LIB_DIR not containing the built libraries
did not break the build. Either way after this commit it will point to
the correct place.

For convenience the _notmuch_config.py from notmuch.bindingconfig is
viewable here
https://gist.github.com/keegancsmith/bdc2f0e44425ad68c8b6a85f8ed0b289

We do not remove the bindingconfig output from notmuch in this commit
since in my local testing that just caused cache invalidations on
notmuch. But can do that as well.